Refactor: reorganise implementation and test files into respective directory strutures that match the challenge id codes
