+++ lib/send_command_to_server.js
@@ -4,6 +4,14 @@ var TDL = require('tdl-client');
 var UserInputAction = require('./runner/user_input_action');
 var utils = require('./runner/utils');
 
+var runner = new TDL.QueryBasedImplementationRunnerBuilder()
+    .setConfig(utils.getRunnerConfig())
+    .withSolutionFor('checkout', require('./solutions/CHK/checkout'))
+    .withSolutionFor('fizz_buzz', require('./solutions/FIZ/fizz_buzz'))
+    .withSolutionFor('hello', require('./solutions/HLO/hello'))
+    .withSolutionFor('sum', require('./solutions/SUM/sum'))
+    .create();
+
 /**
  * ~~~~~~~~~~ Running the system: ~~~~~~~~~~~~~
  *
@@ -45,14 +53,6 @@ var utils = require('./runner/utils');
  *
  **/
 var args = process.argv.slice(2, process.argv.length);
-
-var runner = new TDL.QueryBasedImplementationRunnerBuilder()
-    .setConfig(utils.getRunnerConfig())
-    .withSolutionFor('sum', require('./solutions/sum'))
-    .withSolutionFor('hello', require('./solutions/hello'))
-    .withSolutionFor('fizz_buzz', require('./solutions/fizz_buzz'))
-    .withSolutionFor('checkout', require('./solutions/checkout'))
-    .create();
 
 TDL.ChallengeSession
     .forRunner(runner)
