+++ lib/solutions/CHK/checkout.js
