+++ lib/solutions/SUM/sum.js
