+++ test/solutions/SUM/test-sum.js
@@ -1,4 +1,4 @@
-var sum = require('../../lib/solutions/sum');
+var sum = require('../../../lib/solutions/SUM/sum');
 
 exports['compute sum'] = function (test) {
     test.equal(sum(1, 2), 3);
